Move function to send tokens to user in Python UAT tests common module
