+++ py/tests/integration/test_unsigned_tx.py
@@ -285,25 +285,11 @@ def setup_node_with_tokens(test_settings, node_name):
     return node, common.setup_node_with_tokens(node, test_settings["blocks_to_mine"])
 
 def send_tokens_to_user(user, test_settings, external_api, internal_api):
-    return send_tokens_to_user_(test_settings[user]["pubkey"],
-                                test_settings[user]["amount"],
-                                test_settings[user]["fee"],
-                                external_api,
-                                internal_api)
-
-def send_tokens_to_user_(address, tokens, fee, external_api, internal_api):
-    def get_balance(k):
-        return common.get_account_balance(external_api, internal_api, k).balance
-    bal0 = get_balance(address)
-    spend_tx_obj = SpendTx(
-        recipient_pubkey=address,
-        amount=tokens,
-        fee=fee,
-        ttl=100,
-        payload="sending tokens")
-    internal_api.post_spend_tx(spend_tx_obj)
-    wait(lambda: get_balance(address) == (bal0 + tokens),
-         timeout_seconds=120, sleep_seconds=0.25)
+    return common.send_tokens_to_unchanging_user(test_settings[user]["pubkey"],
+                                                 test_settings[user]["amount"],
+                                                 test_settings[user]["fee"],
+                                                 external_api,
+                                                 internal_api)
 
 def get_unsigned_contract_create(owner, contract, external_api):
     bytecode = read_id_contract(external_api)
